py/requirements.txt
onnxruntime
Pillow
fastapi[standard-no-fastapi-cloud-cli]
numpy